+++ lightning/open_channel.js
@@ -2,6 +2,7 @@ const asyncAuto = require('async/auto');
 const {returnResult} = require('asyncjs-util');
 
 const addPeer = require('./add_peer');
+const {getFundingShim} = require('./../wallet');
 
 const defaultMinConfs = 1;
 const defaultMinHtlcMtokens = '1';
@@ -15,9 +16,19 @@ const minChannelTokens = 20000;
 
   LND 0.8.2 and below do not support `cooperative_close_address`
 
+  External funding requires LND compiled with `walletrpc` build tag
+
   {
     [chain_fee_tokens_per_vbyte]: <Chain Fee Tokens Per VByte Number>
     [cooperative_close_address]: <Restrict Cooperative Close To Address String>
+    [external_funding]: {
+      id: <Pending Channel Id Hex String>
+      key_family: <Funding Key HD Family Number>
+      key_index: <Funding Key HD Index Number>
+      partner_key: <Partner Multisig Public Key Hex String>
+      transaction: <Funding Transaction Hex String>
+      vout: <Funding Transaction Output Index Number>
+    }
     [give_tokens]: <Tokens to Gift To Partner Number> // Defaults to zero
     [is_private]: <Channel is Private Bool> // Defaults to false
     lnd: <Authenticated LND gRPC API Object>
@@ -73,6 +84,25 @@ module.exports = (args, cbk) => {
         cbk);
       }],
 
+      // Get the funding shim for external funding
+      getShim: ['validate', ({}, cbk) => {
+        // Exit early when there is no external funding
+        if (!args.external_funding) {
+          return cbk();
+        }
+
+        return getFundingShim({
+          id: args.external_funding.id,
+          key_family: args.external_funding.key_family,
+          key_index: args.external_funding.key_index,
+          lnd: args.lnd,
+          partner_key: args.external_funding.partner_key,
+          transaction: args.external_funding.transaction,
+          vout: args.external_funding.vout,
+        },
+        cbk);
+      }],
+
       // Determine the minimum confs for spend utxos
       minConfs: ['validate', ({}, cbk) => {
         if (args.min_confirmations === undefined) {
@@ -83,8 +113,17 @@ module.exports = (args, cbk) => {
       }],
 
       // Open the channel
-      openChannel: ['connect', 'minConfs', ({minConfs}, cbk) => {
+      openChannel: [
+        'connect',
+        'getShim',
+        'minConfs',
+        ({getShim, minConfs}, cbk) =>
+      {
         let isAnnounced = false;
+
+        if (!!getShim && getShim.amt !== args.local_tokens.toString()) {
+          return cbk([400, 'ExpectedFundingShimUtxoToMatchChannelCapacity']);
+        }
 
         const options = {
           local_funding_amount: args.local_tokens,
@@ -102,6 +141,10 @@ module.exports = (args, cbk) => {
 
         if (!!args.cooperative_close_address) {
           options.close_address = args.cooperative_close_address;
+        }
+
+        if (!!args.external_funding) {
+          options.funding_shim = {shim: getShim};
         }
 
         if (!!args.give_tokens) {