+++ wallet/index.js
@@ -1,5 +1,11 @@
 const broadcastChainTransaction = require('./broadcast_chain_transaction');
 const getChainFeeRate = require('./get_chain_fee_rate');
+const getFundingShim = require('./get_funding_shim');
 const getPublicKey = require('./get_public_key');
 
-module.exports = {broadcastChainTransaction, getChainFeeRate, getPublicKey};
+module.exports = {
+  broadcastChainTransaction,
+  getChainFeeRate,
+  getFundingShim,
+  getPublicKey,
+};